pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cafa6ml"
version = "0.1.0"
dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "scikit-learn",
  "tqdm",
  "pyyaml",
  "torch"
]

[tool.setuptools]
package-dir = {"" = "src"}
